// src/app/modules/Talent/Talent.interface.ts
export interface ITalentOverview {
  companyName: string;
  talent: string;
  function: string;
  identifiedSkillsGaps: string;
  priorityForFillingGaps: string;
  trainingAndDevelopmentNeeds: string;
  marketTrendsAffectingWorkforce: string;
  talentShortageRisks: string;
  regulatoryChanges: string;
  otherTalentRelatedRisks: string;
  summary: string;
  actionItems: string;
  createdAt?: Date;
  updatedAt?: Date;
}